+++ agileboot-domain/src/main/java/com/agileboot/domain/system/notice/NoticeApplicationService.java
@@ -58,7 +58,7 @@ public class NoticeApplicationService {
         noticeModel.updateById();
     }
 
-    public void deleteNotice(BulkOperationCommand<Long> deleteCommand) {
+    public void deleteNotice(BulkOperationCommand<Integer> deleteCommand) {
         noticeService.removeBatchByIds(deleteCommand.getIds());
     }
